Complete Exercise 2.15. Edit Exercise 2.14

+++ part-2/phonebook/src/components/ContactList.jsx
@@ -3,9 +3,11 @@ import personsService from '../services/persons'
 const ContactList = ({persons, setPersons}) => {
   
   const deleteContact = (person) => {
-    personsService.deleteContact(person);
-    const newPersonsArray = persons.filter(contact => contact.id !== person.id);
-    setPersons(newPersonsArray);
+    if (window.confirm(`delete ${person.name}?`)) {
+      personsService.deleteContact(person);
+      const newPersonsArray = persons.filter(contact => contact.id !== person.id);
+      setPersons(newPersonsArray);
+    }
   }
 
   return (
